goals/requirements.txt
asgiref==3.7.2
autopep8==1.5.7
awscli==1.29.75
botocore==1.31.75
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
dj-database-url==0.5.0
Django==4.2.6
django-admin-material-dashboard==1.0.10
django-adminlte2==0.4.3
django-allauth==0.57.0
django-braces==1.15.0
django-extensions==3.2.3
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.16
filelock==3.13.1
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.5.0
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
Pillow==10.1.0
platformdirs==3.11.0
pyasn1==0.5.0
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.7.0
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.8.0
tzdata==2023.3
Unipath==1.1
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.6
whitenoise==5.3.0
wrapt==1.15.0
